+++ CHANGELOG.md
@@ -25,6 +25,8 @@ This is used for recording Invader's changes. This changelog is based on
   all elements from the array
 - invader-edit-qt: The "Add" button now appends elements to the bottom of the
   array similar to what Guerilla does
+- invader-sound: Specifying the format is now required whenever sound data is
+  being put into a map
 
 ### Fixed
 - invader-edit-qt: Fixed a performance issue on Windows